requirements.txt
texify
doclayout-yolo
pdf2image
ollama
pylatexenc
pytesseract
opencv-python-headless
torch 
torchvision
torchaudio
chromadb
transformers
adapters
Pillow
semantic-text-splitter